requirements.txt
streamlit==1.25.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
